#2113: Use graphql-api to fetch plain articles
When doing SSR we use apollos `renderToStringWithData`. This waits for
graphql requests to finish before returning. When using regular fetch
calls only the initial html will be rendered.

If one wanted to use fetch to aquire data it would have to be done
before the `renderToStringWithData` function is called, and to be used
from props. (IE: Use a class component or rewrite the
`loadGetInitialProps` in defaultRoute.js)